twitter-sentiment-processor/demos/py-demo/processor/requirements.txt
azure-ai-textanalytics >= 5.1.0b2
dapr-ext-grpc >= 0.8.0a0
dapr >= 0.8.0a0